MAUI example for Super Resolution.
Uses RealESRGAN which is a signficantly larger model but has significantly better output than the simple pytorch model we have been using. Pre/post processing is different though so including script to add that.

This could be a template for building out example models to keep things modular and simple - i.e. example app + script that acquires model and defines the full PrePostProcessor pipeline locally. The predefined helpers in extensions are somewhat model specific anyway and it's a little indirect to use a pipeline defined there.