Add migx bert squad quant example (#441)
* Initial Commit of bert quantization with squad v1.1 and v2.0

Added additional pieces with argparse to select which version of squad we want to test, batch size, sequence length and some other useful things.

* Update output of script to show MIGraphX

* Add Latency measurement for inferences

* Add additional input options for debugging as well as io_binding for runs

Seeing stall on larger batch sizes. Adding flags for debugging.

* Fix path for vocab.txt

* Add additional flags for resnet50 int8 run

- Allow for mixed precsion via fp16 flag
-Allow for variable batch
-Allow for variable calibrate data size

* update compute_data()

* Fix error with json serialization of calibration data

* Update script to enforce calibration cache name and change flags based on version

* Use model path directly

* Remove usage of path for e2e bert model

* Load onnx model before calibration begins

* Remove false from onnx load

* Remove strided data reader

Running into an issue with shapes with the calibration tools if I break up this calibration read. This needs a large amount of memory to create the histogram.

* Arg changes for bert script

* Fix merge conflicts

* use ort_session instead of session

* Fix error message with sample size

* Additional changes to handle various model inputs

Useful if we want to use another falvor of bert for now.

TODO need to handle/fix some of the input/output arg maps vs the input data vs model input/outputs

* Add sequence length to mxr file output naming

* Add query length input parameter

Another knob to tune/play with in perf runs. Right now just allow this to be default.

* Modify script to add padding for now due to varying batch size used to handle features in each example

Our MIGraphX EP requires a recompile of the model if we constantly change the input dimensions or batch size of the parameters. Without this we actually cause a slowdown with the larger batch size runs as we tend to go above the feature index.

A workaround is to ensure that batch size stays constant as we feed data into the model we're testing to get inference timing and accuracy results via repeating the same sample until we have enough data for a proper batch size.

* Set toggle for save_load of models

* Add EP option and gate out model run with save/load

* Fix querry_length to query_len

* Add option for calibration EP data selection

* Additional Fixes for save_load and adding CPU EP option

* Only print quantizer info for int8 runs

* Update README

---------

Co-authored-by: Ted Themistokleous <tedthemistokleous@amd.com>